# burrito/apps/statistic/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an app to provide with a statistic information",
        "title": "BurritoStatistic",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/statistic/activity_summary": {
            "get": {
                "description": "get period statistic",
                "produces": [
                    "application/json"
                ],
                "summary": "get period statistic",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ActivitySummaryModel"
                        }
                    }
                }
            }
        },
        "/statistic/faculty": {
            "get": {
                "description": "get faculty statistic",
                "produces": [
                    "application/json"
                ],
                "summary": "get faculty statistic",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FacultyStatisticModel"
                        }
                    }
                }
            }
        },
        "/statistic/period": {
            "post": {
                "description": "get general statistic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get general statistic",
                "parameters": [
                    {
                        "description": "PeriodStatisticRequest",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PeriodStatisticRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PeriodStatisticModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ActivitySummaryModel": {
            "type": "object",
            "properties": {
                "average_process_time": {
                    "type": "number"
                },
                "tickets_processed": {
                    "type": "integer"
                },
                "users_registered": {
                    "type": "integer"
                }
            }
        },
        "models.FacultyData": {
            "type": "object",
            "properties": {
                "created_tickets_percent": {
                    "type": "number"
                },
                "faculty_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "registered_users": {
                    "type": "integer"
                }
            }
        },
        "models.FacultyStatisticModel": {
            "type": "object",
            "properties": {
                "faculties_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FacultyData"
                    }
                }
            }
        },
        "models.PeriodStatisticModel": {
            "type": "object",
            "properties": {
                "scopes": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "date": {
                                "type": "string"
                            },
                            "scope": {
                                "type": "integer"
                            },
                            "tickets_count": {
                                "type": "integer"
                            }
                        }
                    }
                },
                "statuses": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "date": {
                                "type": "string"
                            },
                            "status_id": {
                                "type": "integer"
                            },
                            "status_name": {
                                "type": "string"
                            },
                            "tickets_count": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
        "models.PeriodStatisticRequest": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        }
    }
}