# burrito/apps/statistic/models/period_statistic_model.go
package models

type PeriodStatisticRequest struct {
	From string `json:"from"`
	To   string `json:"to"`
}

type PeriodStatisticModel struct {
	ByStatuses []struct {
		Date         string `json:"date"`
		StatusID     int    `json:"status_id"`
		StatusName   string `json:"status_name"`
		TicketsCount int64  `json:"tickets_count"`
	} `json:"statuses"`

	ByScopes []struct {
		Date         string `json:"date"`
		Scope        int    `json:"scope"`
		TicketsCount int64  `json:"tickets_count"`
	} `json:"scopes"`
}